# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Marco Api Document",
        "description": "This is the API documentation for our project.",
        "contact": {
            "email": "alibabaeian670@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/admin/user": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get list of users",
                "description": "Only accessible for authenticated users",
                "operationId": "6d60bab3d69eff694999ea65500674bd",
                "responses": {
                    "200": {
                        "description": "List of users"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/login": {
            "post": {
                "summary": "Login user",
                "description": "Login user with phone number and code",
                "operationId": "de45964bd236f87e5860ad7e58ec540f",
                "parameters": [
                    {
                        "name": "phone",
                        "in": "query",
                        "description": "Login with phone number",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "code",
                        "in": "query",
                        "description": "Verification code",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful login",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string"
                                        },
                                        "client_id": {
                                            "type": "integer"
                                        },
                                        "client_secret": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "203": {
                        "description": "Verification code error"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "passport": {
                "type": "http",
                "description": "Use Passport bearer token to access protected endpoints",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "User",
            "description": "User"
        }
    ]
}